pyproject.toml
[project]
name = "studybuddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ambreen874", email = "ambreensharif55@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
]

[project.scripts]
studybuddy = "studybuddy:main"
studybuddy-run = "studybuddy.agent_study:studybuddy_agent"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
